Display (broken) data-bound context menu in ControlCatalog.
To demonstrate #2325. Moved `MenuPageViewModel` and `MenuItemViewModel` to new `ViewModels` directory so the latter can be reused for `ContextMenuPageViewModel`.
